+++ src/App.tsx
@@ -1,44 +1,16 @@
-import { useEffect, useState } from 'react';
-import { dummyData } from './data/todos';
 import AddTodoForm from './components/AddTodoForm';
 import TodoList from './components/TodoList';
 import TodoSummary from './components/TodoSummary';
-import { Todo } from './types/todo';
+import useTodos from './hooks/useTodos';
 
 function App() {
-  const [todos, setTodos] = useState(() => {
-    const saveTodos: Todo[] = JSON.parse(localStorage.getItem('todos') || '[]');
-    return saveTodos.length > 0 ? saveTodos : dummyData;
-  });
-
-  useEffect(() => {
-    localStorage.setItem('todos', JSON.stringify(todos));
-  }, [todos]);
-
-  function setTodoCompleted(id: number, completed: boolean) {
-    setTodos((prevTodos) =>
-      prevTodos.map((todo) => (todo.id === id ? { ...todo, completed } : todo))
-    );
-  }
-
-  function addTodo(title: string) {
-    setTodos((prevTodos) => [
-      {
-        id: Date.now(),
-        title,
-        completed: false,
-      },
-      ...prevTodos,
-    ]);
-  }
-
-  function deleteTodo(id: number) {
-    setTodos((prevTodos) => prevTodos.filter((todo) => todo.id !== id));
-  }
-
-  function deleteAllCompletedTodos() {
-    setTodos((prevTodos) => prevTodos.filter((todo) => !todo.completed));
-  }
+  const {
+    todos,
+    addTodo,
+    setTodoCompleted,
+    deleteAllCompletedTodos,
+    deleteTodo,
+  } = useTodos();
 
   return (
     <main className='py-10 h-screen space-y-5 overflow-y-auto'>
